msm: camera: cpp: Vote AHB with clock enabled and no mem retention
Vote for AHB bandwidth after clocks are enabled. For bandwidth
voting clocks need to be on. Set the memory retention flags for
certain hardware clocks to save power, memory does not need to
be retained for these clocks since the driver reinitializes the
jpeg hardware whenever the clocks are turned on.

CRs-Fixed: 1044777 1044373
Signed-off-by: Rajakumar Govindaram <rajakuma@codeaurora.org>
Change-Id: I51ec59b2940b146e5c1a24e21ee06e6332e23aff
Signed-off-by: Gautham Mayyuri <gmayyuri@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>