sched: walt: Optimize task_util()
task_util() for WALT is currently defined as

p->ravg.demand / (sched_ravg_window >> SCHED_CAPACITY_SHIFT);

This math is required to scale the task demand to 1024 scale.
task_util() is used many times in task placement. So the calls
to this can be optimized by caching the scaled value when task
demand is calculated.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Change-Id: If02fdc0d087f669ed6a1a73e91d7282d08275dc1
Signed-off-by: TogoFire <italomellopereira@gmail.com>