media: cec: check for non-OK/NACK conditions while claiming a LA
During the configuration phase of a CEC adapter it is trying to claim a
free logical address by polling.

Change-Id: a3837fd16026ef64b398f6ce37f4303c6a8e6cb6
However, the code doesn't check if there were errors other than OK or NACK,
those are just treated as if the poll was NACKed.

Instead check for such errors and retry the poll. And if the problem persists
then don't claim this LA since there is something weird going on.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Signed-off-by: TogoFire <togofire@mailfence.com>