kobject_uevent: Allocate environment buffer on the stack
The environment buffer isn't very big; when it's allocated on the stack,
kobject_uevent_env's stack frame size increases to just over 2 KiB,
which is safe considering that we have a 16 KiB stack.

Allocate the environment buffer on the stack instead of using the slab
allocator in order to improve performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I5321e40968e37182642f8ab18dca8a30b436f587
Signed-off-by: Lau <laststandrighthere@gmail.com>