wireguard: allowedips: initialize list head in selftest
The randomized trie tests weren't initializing the dummy peer list head,
resulting in a NULL pointer dereference when used. Fix this by
initializing it in the randomized trie test, just like we do for the
static unit test.

While we're at it, all of the other strings like this have the word
"self-test", so add it to the missing place here.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I7266258a4d986b612e5c61d2350cdd59a7f58f0e
Signed-off-by: TogoFire <togofire@mailfence.com>