wifi: ath5k: fix an off by one check in ath5k_eeprom_read_freq_list()
[ Upstream commit 4c856ee12df85aabd437c3836ed9f68d94268358 ]

This loop checks that i < max at the start of loop but then it does
i++ which could put it past the end of the array.  It's harmless to
check again and prevent a potential out of bounds.

Fixes: 1048643ea94d ("ath5k: Clean up eeprom parsing and add missing calibration data")
Signed-off-by: Dan Carpenter <error27@gmail.com>
Change-Id: I9f2b31a7ae292a668cd024faefc66cae7747af7d
Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/Y+D9hPQrHfWBJhXz@kili
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit 6b8ef48c6f4493aa9ec6b0d0660c9d87374d9198)
Signed-off-by: Togo 77 <togofire@mailfence.com>