mips: Move initrd_start check after initrd address sanitisation.
[ Upstream commit 4897a898a216058dec55e5e5902534e6e224fcdf ]

PAGE_OFFSET is technically a virtual address so when checking the value of
initrd_start against it we should make sure that it has been sanitised from
the values passed by the bootloader. Without this change, even with a bootloader
that passes correct addresses for an initrd, we are failing to load it on MT7621
boards, for example.

Signed-off-by: Liviu Dudau <liviu@dudau.co.uk>
Change-Id: I24d54b01b570eba72dc2fa14b13ea5d7136bd295
Signed-off-by: Thomas Bogendoerfer <tsbogend@alpha.franken.de>
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit 3349fdfb44249618f807d8a53bc22fedb0857381)
Signed-off-by: TogoFire <togofire@mailfence.com>