bus: mhi: misc: Add check for dev_rp if it is iommu range or not
er_ctxt->rp pointer is updated by MDM which is untrusted to HLOS,
it could be arbitrary value.

If there is security issue on MDM, and updated pointer which is not
align then driver will never come out of loop where checking against
dev_rp != rp.

So added check to make sure it is in the buffer range & aligned to 128bit.

Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>
Change-Id: I6730406c125037791da7c6c11e9af60ab53e4b2f
Signed-off-by: Paras Sharma <quic_parass@quicinc.com>
Signed-off-by: TogoFire <togofire@mailfence.com>